+++ common/ASC.Core.Common/Notify/Telegram/Dao/TelegramUser.cs
@@ -1,6 +1,6 @@
 ﻿/*
  *
- * (c) Copyright Ascensio System Limited 2010-2023
+ * (c) Copyright Ascensio System Limited 2010-2020
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,10 @@
 
 
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace ASC.Core.Common.Notify.Telegram
 {
@@ -23,6 +27,6 @@ namespace ASC.Core.Common.Notify.Telegram
     {
         public Guid PortalUserId { get; set; }
         public int TenantId { get; set; }
-        public long TelegramId { get; set; }
+        public int TelegramId { get; set; }
     }
 }
